+++ src/tasks/Task_27.java
@@ -3,9 +3,9 @@ package tasks;
 public class Task_27 {
 
     public static void main(String[] args) {
-        double a = 0;
-        double b = 10;
-        double c = 15;
+        int a = 0;
+        int b = 10;
+        int c = 15;
 
         if (a - b == b - c) {
             System.out.println(true);
