+++ src/tasks/Task_33.java
@@ -3,10 +3,10 @@ package tasks;
 public class Task_33 {
 
     public static void main(String[] args) {
-        double a = 20;
-        double b = 14;
-        double c = 50;
-        double d = 1;
+        int a = 20;
+        int b = 14;
+        int c = 50;
+        int d = 1;
 
         if (a == 1 || b == 1 || c == 1 || d == 1) {
             System.out.println(true);
