+++ src/tasks/Task_36.java
@@ -3,10 +3,10 @@ package tasks;
 public class Task_36 {
 
     /*public static void main(String[] args) {
-        double a = 1;
-        double b = 3;
-        double c = 6;
-        double d = 6;
+        int a = 1;
+        int b = 3;
+        int c = 6;
+        int d = 6;
         if ((a % 2 != 0 && b % 2 != 0) || (a % 2 != 0 && c % 2 != 0)|| (a % 2 != 0 && d % 2 != 0)|| (b % 2 != 0 && c % 2 != 0)|| (b % 2 != 0 && d % 2 != 0)|| (c % 2 != 0 && d % 2 != 0)) {
             System.out.println(1);
         } else {
@@ -14,10 +14,10 @@ public class Task_36 {
         }
     }*/
     public static void main(String[] args) {
-        double a = 1;
-        double b = 3;
-        double c = 6;
-        double d = 6;
+        int a = 1;
+        int b = 3;
+        int c = 6;
+        int d = 6;
         if (a % 2 != 0 && b % 2 != 0){
             System.out.println(1);
         } else if (a % 2 != 0 && c % 2 != 0){
